# online_products_apispec.yaml
---
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
  "/all-products":
    get:
      tags:
        - products
      summary: Get All Products
      operationId: get_all_products_all_products_get
      parameters:
        - name: internal-api-key
          in: header
          required: true
          schema:
            type: string
            title: Internal-Api-Key
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/product/{product_id}":
    get:
      tags:
        - products
      summary: Get Product By Id
      operationId: get_product_by_id_product__product_id__get
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
            title: Product Id
        - name: internal-api-key
          in: header
          required: true
          schema:
            type: string
            title: Internal-Api-Key
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/products-by-category/{category_id}":
    get:
      tags:
        - products
      summary: Get Products By Category
      operationId: get_products_by_category_products_by_category__category_id__get
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: integer
            title: Category Id
        - name: internal-api-key
          in: header
          required: true
          schema:
            type: string
            title: Internal-Api-Key
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/create-product/category/{category_id}":
    post:
      tags:
        - products
      summary: Create Product
      operationId: create_product_create_product_category__category_id__post
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: integer
            title: Category Id
        - name: internal-api-key
          in: header
          required: true
          schema:
            type: string
            title: Internal-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ProductRequest"
      responses:
        "201":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/update-product/{product_id}":
    put:
      tags:
        - products
      summary: Update Product
      operationId: update_product_update_product__product_id__put
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
            title: Product Id
        - name: internal-api-key
          in: header
          required: true
          schema:
            type: string
            title: Internal-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ProductRequest"
      responses:
        "202":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/delete-product/{product_id}":
    delete:
      tags:
        - products
      summary: Delete Product
      operationId: delete_product_delete_product__product_id__delete
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
            title: Product Id
        - name: internal-api-key
          in: header
          required: true
          schema:
            type: string
            title: Internal-Api-Key
      responses:
        "204":
          description: Successful Response
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/all-categories":
    get:
      tags:
        - categories
      summary: Get All Categories
      operationId: get_all_categories_all_categories_get
      parameters:
        - name: internal-api-key
          in: header
          required: true
          schema:
            type: string
            title: Internal-Api-Key
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/category/{category_id}":
    get:
      tags:
        - categories
      summary: Get Category By Id
      operationId: get_category_by_id_category__category_id__get
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: integer
            title: Category Id
        - name: internal-api-key
          in: header
          required: true
          schema:
            type: string
            title: Internal-Api-Key
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/create-category":
    post:
      tags:
        - categories
      summary: Create Category
      operationId: create_category_create_category_post
      parameters:
        - name: internal-api-key
          in: header
          required: true
          schema:
            type: string
            title: Internal-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CategoryRequest"
      responses:
        "201":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/update-category":
    put:
      tags:
        - categories
      summary: Update Category
      operationId: update_category_update_category_put
      parameters:
        - name: category_id
          in: query
          required: true
          schema:
            type: integer
            title: Category Id
        - name: internal-api-key
          in: header
          required: true
          schema:
            type: string
            title: Internal-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CategoryRequest"
      responses:
        "202":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
components:
  schemas:
    CategoryRequest:
      properties:
        name:
          type: string
          maxLength: 50
          minLength: 1
          title: Name
        description:
          type: string
          maxLength: 255
          minLength: 6
          title: Description
      type: object
      required:
        - name
        - description
      title: CategoryRequest
    HTTPValidationError:
      properties:
        detail:
          items:
            "$ref": "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ProductRequest:
      properties:
        name:
          type: string
          maxLength: 150
          title: Name
        description:
          type: string
          maxLength: 255
          title: Description
        price:
          type: number
          exclusiveMinimum: 0
          title: Price
        stock:
          type: integer
          exclusiveMinimum: 0
          title: Stock
      type: object
      required:
        - name
        - description
        - price
        - stock
      title: ProductRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError